Enhance User functionality by adding XP attribute and updating UserController to support fetching by user ID. Modify UserService to handle XP updates and introduce error handling for workout creation in WorkoutService.

+++ src/main/java/com/MuscleHead/MuscleHead/Workout/Workout.java
@@ -30,7 +30,7 @@ public class Workout {
 
     @Column(updatable = false, columnDefinition = "TIMESTAMP")
     private Instant date;
-    
+
     private String notes;
     private String workout_name;
 
@@ -46,7 +46,7 @@ public class Workout {
         return workout_id;
     }
 
-    public void setWorkout_id(int workout_id) {
+    public void setWorkout_id(long workout_id) {
         this.workout_id = workout_id;
     }
